envoy/envoy-mysql-app/app/requirements.txt
mysql-connector-python
flask
jsonify

